Logs, styling, typos.
Added a log of sorts to log monster kills. This, in the long run, will prove to be a better and easier system for letting players keep track of things that are happening in-game. Plus, I changed some of the styling so that staring at it won't hurt my eyes too much. And I also fixed some typos I saw.

First commit - just trying out some things. This code is also pretty clunky (see fight.js) but with some tidying up this can be way better.